Extract JSON objects into a separate crate
