+++ ofdb-boundary/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "ofdb-boundary"
+version = "0.0.0" # will stay 0.0.0 until published
+authors = ["slowtec GmbH <post@slowtec.de>"]
+edition = "2018"
+publish = false
+
+[dependencies]
+serde = { version = "1", features = ["derive"] }
+
+[features]
+default = []
+extra-derive = []
