+++ ofdb-boundary/src/lib.rs
@@ -1,0 +1,221 @@
+use serde::{Deserialize, Serialize};
+
+#[rustfmt::skip]
+#[derive(Serialize, Deserialize)]
+#[cfg_attr(feature = "extra-derive", derive(Debug, Clone, PartialEq))]
+pub struct Entry {
+    pub id             : String,
+    pub created        : i64,
+    pub version        : u64,
+    pub title          : String,
+    pub description    : String,
+    pub lat            : f64,
+    pub lng            : f64,
+    pub street         : Option<String>,
+    pub zip            : Option<String>,
+    pub city           : Option<String>,
+    pub country        : Option<String>,
+    pub email          : Option<String>,
+    pub telephone      : Option<String>,
+    pub homepage       : Option<String>,
+    pub categories     : Vec<String>,
+    pub tags           : Vec<String>,
+    pub ratings        : Vec<String>,
+    pub license        : Option<String>,
+    pub image_url      : Option<String>,
+    pub image_link_url : Option<String>,
+}
+
+#[derive(Serialize, Deserialize)]
+#[cfg_attr(feature = "extra-derive", derive(Debug, Clone))]
+pub struct Event {
+    pub id: String,
+    pub title: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub description: Option<String>,
+    pub start: i64,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub end: Option<i64>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub lat: Option<f64>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub lng: Option<f64>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub street: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub zip: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub city: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub country: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub email: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub telephone: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub homepage: Option<String>,
+    pub tags: Vec<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub registration: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub organizer: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub image_url: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub image_link_url: Option<String>,
+}
+
+#[derive(Serialize, Deserialize)]
+#[cfg_attr(feature = "extra-derive", derive(Debug, Clone, Copy))]
+pub struct Coordinate {
+    pub lat: f64,
+    pub lng: f64,
+}
+
+#[derive(Serialize, Deserialize)]
+#[cfg_attr(feature = "extra-derive", derive(Debug, Clone))]
+pub struct User {
+    pub email: String,
+    pub email_confirmed: bool,
+    pub role: UserRole,
+}
+
+#[derive(Serialize, Deserialize)]
+#[cfg_attr(feature = "extra-derive", derive(Debug, Clone, Copy))]
+pub struct RatingValue(i8);
+
+#[derive(Serialize, Deserialize)]
+#[cfg_attr(feature = "extra-derive", derive(Debug, Clone, Copy))]
+pub struct AvgRatingValue(f64);
+
+impl From<f64> for AvgRatingValue {
+    fn from(v: f64) -> Self {
+        Self(v)
+    }
+}
+
+impl From<AvgRatingValue> for f64 {
+    fn from(from: AvgRatingValue) -> Self {
+        from.0
+    }
+}
+
+impl From<i8> for RatingValue {
+    fn from(from: i8) -> Self {
+        Self(from)
+    }
+}
+
+#[derive(Serialize, Deserialize)]
+#[cfg_attr(
+    feature = "extra-derive",
+    derive(Debug, Clone, Copy, PartialEq, Eq, Hash)
+)]
+#[serde(rename_all = "snake_case")]
+pub enum RatingContext {
+    Diversity,
+    Renewable,
+    Fairness,
+    Humanity,
+    Transparency,
+    Solidarity,
+}
+
+#[derive(Serialize, Deserialize)]
+#[cfg_attr(feature = "extra-derive", derive(Debug, Clone))]
+pub struct EntrySearchRatings {
+    pub total: AvgRatingValue,
+    pub diversity: AvgRatingValue,
+    pub fairness: AvgRatingValue,
+    pub humanity: AvgRatingValue,
+    pub renewable: AvgRatingValue,
+    pub solidarity: AvgRatingValue,
+    pub transparency: AvgRatingValue,
+}
+
+#[derive(Serialize, Deserialize)]
+#[cfg_attr(feature = "extra-derive", derive(Debug, Clone))]
+pub struct Comment {
+    pub id: String,
+    pub created: i64,
+    pub text: String,
+}
+
+#[derive(Serialize, Deserialize)]
+#[cfg_attr(feature = "extra-derive", derive(Debug, Clone))]
+pub struct Category {
+    pub id: String,
+    pub name: String,
+}
+
+#[derive(Serialize, Deserialize)]
+#[cfg_attr(feature = "extra-derive", derive(Debug, Clone))]
+pub struct EntrySearchResult {
+    pub id: String,
+    pub lat: f64,
+    pub lng: f64,
+    pub title: String,
+    pub description: String,
+    pub categories: Vec<String>,
+    pub tags: Vec<String>,
+    pub ratings: EntrySearchRatings,
+}
+
+#[derive(Serialize, Deserialize)]
+#[cfg_attr(feature = "extra-derive", derive(Debug, Clone))]
+pub struct SearchResponse {
+    pub visible: Vec<EntrySearchResult>,
+    pub invisible: Vec<EntrySearchResult>,
+}
+
+#[derive(Serialize, Deserialize)]
+#[cfg_attr(
+    feature = "extra-derive",
+    derive(Debug, Clone, Copy, PartialEq, Eq, Hash)
+)]
+#[serde(rename_all = "lowercase")]
+pub enum UserRole {
+    Guest,
+    User,
+    Scout,
+    Admin,
+}
+
+#[derive(Serialize, Deserialize)]
+#[cfg_attr(feature = "extra-derive", derive(Debug, Clone))]
+pub struct BboxSubscription {
+    pub id: String,
+    pub south_west_lat: f64,
+    pub south_west_lng: f64,
+    pub north_east_lat: f64,
+    pub north_east_lng: f64,
+}
+
+#[derive(Serialize, Deserialize)]
+#[cfg_attr(feature = "extra-derive", derive(Debug, Clone))]
+pub struct RequestPasswordReset {
+    pub email: String,
+}
+
+#[derive(Serialize, Deserialize)]
+#[cfg_attr(feature = "extra-derive", derive(Debug, Clone))]
+pub struct ResetPassword {
+    pub token: String,
+    pub new_password: String,
+}
+
+#[derive(Serialize, Deserialize)]
+#[cfg_attr(feature = "extra-derive", derive(Debug, Clone))]
+pub struct TagFrequency(pub String, pub u64);
+
+#[derive(Serialize, Deserialize)]
+#[cfg_attr(feature = "extra-derive", derive(Debug, Clone))]
+pub struct Rating {
+    pub id: String,
+    pub title: String,
+    pub created: i64,
+    pub value: RatingValue,
+    pub context: RatingContext,
+    pub comments: Vec<Comment>,
+    pub source: String,
+}
